Version 0.3.0-SNAPSHOT Start
- split `SparkRunnable` into `SparkRunnable` and `SparkApp`
