+++ docs/publishing.md
@@ -1,16 +1,19 @@
 # Publishing the Project to Sonatype
 
+
 `sbt publishSigned`
 
 `sbt publishLocalSigned`
 
 `sbt release`
 
+
 ## External References
 
 - [OSSRH Guide](https://central.sonatype.org/pages/ossrh-guide.html)
 - [Working with PGP Signatures](https://central.sonatype.org/pages/working-with-pgp-signatures.html)
 - [SBT Using Sonatype](https://www.scala-sbt.org/0.13/docs/Using-Sonatype.html)
+
 
 ## Other References
