+++ src/test/resources/MockApp/application.conf
@@ -1,0 +1,6 @@
+# Resource file for the SparkRunnableSpec test
+
+MockApp.whoami = "./src/test/resources/MockApp/application.conf"
+MockApp.some.list = ["a", "b", "c"]
+MockApp.file.application.conf: true
+
