--- src/test/resources/MockRunnable/application.conf
@@ -1,6 +1,0 @@
-# Resource file for the SparkRunnableSpec test
-
-MockRunnable.whoami = "./src/test/resources/MockRunnable/application.conf"
-MockRunnable.some.list = ["a", "b", "c"]
-MockRunnable.file.application.conf: true
-
