+++ src/test/resources/reference.conf
@@ -5,7 +5,7 @@ schemafactory: {
 }
 
 # Configuration section for the SparkRunnableSpec test
-MockRunnable: {
+MockApp: {
   whoami: "./src/test/resources/reference.conf"
   reference: "reference"
   classpath.application.conf: false
