+++ src/test/scala/org/tupol/spark/SparkAppSpec.scala
@@ -8,10 +8,10 @@ import org.scalatest.{ FunSuite, Matchers }
 
 import scala.util.Try
 
-class SparkRunnableSpec extends FunSuite with Matchers with SharedSparkSession {
+class SparkAppSpec extends FunSuite with Matchers with SharedSparkSession {
 
   val filesArg = Seq(
-    new File("src/test/resources/MockRunnable/application.conf").getAbsolutePath)
+    new File("src/test/resources/MockApp/application.conf").getAbsolutePath)
 
   override def sparkConfig: Map[String, String] = {
     // Add the comma separated configuration files to the files property.
@@ -22,12 +22,12 @@ class SparkRunnableSpec extends FunSuite with Matchers with SharedSparkSession {
   }
 
   test(
-    """SparkRunnable.applicationConfiguration loads first the app params then defaults to application.conf file,
+    """SparkApp.applicationConfiguration loads first the app params then defaults to application.conf file,
       |then to the application.conf in the classpath and then to reference.conf""".stripMargin) {
 
-      val conf = MockRunnable.applicationConfiguration(spark, Array(
-        "MockRunnable.whoami=\"app.param\"",
-        "MockRunnable.param=\"param\""))
+      val conf = MockApp$.applicationConfiguration(spark, Array(
+        "MockApp.whoami=\"app.param\"",
+        "MockApp.param=\"param\""))
       conf.getString("param") shouldBe "param"
       conf.getString("whoami") shouldBe "app.param"
       conf.getStringList("some.list").toArray shouldBe Seq("a", "b", "c")
@@ -35,17 +35,17 @@ class SparkRunnableSpec extends FunSuite with Matchers with SharedSparkSession {
       conf.getBoolean("file.application.conf") shouldBe true
     }
 
-  test("SparkRunnable.applicationConfiguration loads first the application.conf then defaults to reference.conf") {
-    val conf = MockRunnable.applicationConfiguration(spark, Array(
-      "MockRunnable.param=\"param\""))
+  test("SparkApp.applicationConfiguration loads first the application.conf then defaults to reference.conf") {
+    val conf = MockApp$.applicationConfiguration(spark, Array(
+      "MockApp.param=\"param\""))
     conf.getString("param") shouldBe "param"
-    conf.getString("whoami") shouldBe "./src/test/resources/MockRunnable/application.conf"
+    conf.getString("whoami") shouldBe "./src/test/resources/MockApp/application.conf"
     conf.getStringList("some.list").toArray shouldBe Seq("a", "b", "c")
     conf.getString("reference") shouldBe "reference"
     conf.getBoolean("file.application.conf") shouldBe true
   }
 
-  object MockRunnable extends SparkRunnable[String, Unit] {
+  object MockApp$ extends SparkApp[String, Unit] {
 
     def buildConfig(config: Config): Try[String] = Try("Hello")
